Checking arguments to launch tools

+++ Segmentation/minUnderMask.m
@@ -2,7 +2,9 @@ function minValue = minUnderMask(img, mask)
 
 logicalMaskLinear = reshape(logical(mask), 1, []);
 imgLinear = reshape(img, 1, []);
-linearCombined = sort(unique(imgLinear .* logicalMaskLinear));
+% imgType = class(imgLinear);
+% maskLinear = typecast(logicalMaskLinear, imgType);
+linearCombined = sort(unique(imgLinear .* double(logicalMaskLinear)));
 if max(linearCombined) == 0
     minValue = 0;
 else
